notemaster/requirements.txt
amqp==2.4.1
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.8.0
billiard==3.5.0.5
celery==4.2.1
certifi==2019.11.28
chardet==3.0.4
cliapp==1.20170827
Click==7.0
cmdtest==0.32
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
deluge==1.3.15
Django==1.11.26
django-allauth==0.40.0
django-easy-pdf==0.1.1
django-rest-auth==0.9.5
django-rest-swagger==2.0.7
djangorestframework==3.9.4
enum34==1.1.6
Flask==1.1.1
html5lib==1.0.1
hyperlink==17.3.1
idna==2.8
incremental==16.10.1
ipaddress==1.0.17
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.3
kombu==4.3.0
Markdown==2.6.9
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
oauthlib==3.1.0
openapi-codec==1.3.2
Parsley==1.2
Pillow==6.2.1
protobuf==3.9.1
psutil==5.4.2
psycopg2==2.7.4
psycopg2-binary==2.8.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pygobject==3.26.1
pyliblzma==0.5.3
PyMySQL==0.9.3
pyOpenSSL==17.5.0
pyparted==3.11.1
pypaystack==1.19
PyPDF2==1.26.0
pypng==0.0.20
python-apt==1.6.3
python-dotenv==0.10.3
python-libtorrent==1.1.5
python-openid==2.2.5
python-xlib==0.20
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.5.32
requests==2.22.0
requests-oauthlib==1.3.0
selenium==3.141.0
service-identity==16.0.0
setproctitle==1.1.10
simplejson==3.17.0
six==1.13.0
soupsieve==1.9.3
ttystatus==0.38
Twisted==17.9.0
txsocksx==0.0.0
typing==3.7.4.1
uritemplate==3.0.0
urllib3==1.25.7
vine==1.2.0
virtualenv==16.7.4
webencodings==0.5.1
Werkzeug==0.15.5
xhtml2pdf==0.2.3
zope.interface==4.3.2
